Handle deleted option in where clauses
Work around the lack of a property to specify deleted values in where
only functions by using a reserved keyword at the top level of the where
clause that is always removed ('$deleted'). This allows these functions
to normally omit deleted items, but like their query counterparts,
include them if the value is set to true.